requirements.txt
# Core dependencies
fastapi
uvicorn
python-multipart
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-dateutil

# Database
pymongo
python-dotenv
certifi>=2023.7.22

# PDF processing
PyPDF2

# AI and ML
transformers
sentence-transformers
faiss-cpu

# LangChain
langchain
langchain-core
langchain-google-genai
google-ai-generativelanguage
langchain-community

pdf2image 
pytesseract 
pillow
